added delegate, enum class; fixed some bugs .

+++ paf/src/pafcore/Enumerator.cpp
@@ -5,7 +5,7 @@
 
 BEGIN_PAFCORE
 
-Enumerator::Enumerator(const char* name, Attributes* attributes, Type* type, size_t value)
+Enumerator::Enumerator(const char* name, Attributes* attributes, Type* type, int value)
 : Metadata(name, attributes)
 {
 	m_type = type;
